LayoutTests/media/modern-media-controls/mute-support/mute-support-muted.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<video src="../../content/test.mp4" style="width: 320px; height: 240px;" muted></video>
<div id="shadow"></div>
<script type="text/javascript">

description("Testing the <code>MuteSupport</code> behavior by using the muted HTML attribute.");

const shadowRoot = document.querySelector("div#shadow").attachShadow({ mode: "open" });
const media = document.querySelector("video");
const mediaController = createControls(shadowRoot, media, null);

shouldBeTrue("mediaController.controls.muteButton.muted");

debug("");
shadowRoot.host.remove();
media.remove();

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
